Add fit values "horizontal" and "vertical"
The fit attributes accepted "inside" and "outside", which determines
whether the fit must be horizontal or vertical depending on the aspect
ratios of the image and the component.

But sometimes, we may want to force the fit horizontally or vertically.

Therefore, add two values to the "fit" attribute, "horizontal" and
"vertical"