Change license from LGPL to MIT
I chose the LGPL license because I wanted the changes made to the
library to be redistributed to users (without forcing the whole
application using it to become GPL).

However, the license GNU/LGPL says (§4.d.0):

> Convey the Minimal Corresponding Source under the terms of this
> License, and the Corresponding Application Code in a form suitable
> for, and under terms that permit, the user to recombine or relink the
> Application with a modified version of the Linked Version to produce a
> modified Combined Work

For an Android application, this is not practical: the library is
combined with the main project to produce an APK. It is not (easily)
possible to replace only the library part with a modified version.

Therefore, I decided to use a more permissive license, MIT.